requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.9

# 机器学习和图像处理
keras-facenet==0.3.2
mtcnn==0.1.1
tensorflow==2.15.0
numpy==1.24.3
scipy==1.11.4
Pillow==10.1.0
opencv-python-headless==4.8.1.78

# 加密
cryptography==41.0.7

# 环境变量管理
python-dotenv==1.0.0

# 代码质量和测试
flake8==6.1.0
pytest==7.4.3
httpx==0.25.2

# 数据版本控制
dvc==3.30.1
dvc-s3==2.23.0